Add files via upload
The magnetic north project is moving forward.  The C++ (arduino) is long, but just simple repetition.  The wiring is similar;  lots of wires and resistors, but just a lot of the same.  There are 40 detectors, but they are all the same.  This applies to both hardware and software.  The database interface (Python) is the same as from the other project but simple.  There is just the one input;  which LDR is stimulated.  The time code is generated in Python.  A database row is just unix time and the specific LDR, recorded as its position.  The overall objective is to record magnetic north every ten minutes and let it run until I come with a more accurate measurement.  I have ideas about that.  The present platform is defensible and robust, but very much not precise.
And this is relative position.
The narrative is really my notes.