latest data as Excel file with graph
This project is under continual development.  But data is coming in and looks interesting.  The raw Excel data is included.  It is edited only to the point of having the error codes taken out.  The python variable is set at 767676 (or something obvious) and if no reading is taken (Like the laser is not turning on,  maybe an LDR is not responding etc) then the default variable indicates a problem.