This is September 2024 update
I want to do a 'system upgrade' for this project.  Maybe the precision of magnetic north measurement could be improved by evaluation of the behavior of a gently swinging needle.    And,   the original platform searched for the Boolean positive I/O with 33 'if' statements.  Pretty messy.  The arduino C++ program is download to the GitHub site.. and thanks