I am trying to create an upgrade to the project.
The detector could be described as a ‘quantum detector’  It detects the movement of magnetic north in quantum increments about two minutes apart  (where there are 60 minutes in a degree.  They are about two millimeters apart.)
A realistic upgrade would be to evaluate magnetic north from a careful evaluation of the behavior of a gently swinging needle.  For Example:  the magnetic needle is swinging back and forth (should be simple harmonic motion)  perhaps with an amplitude of 40 centimeters.  The behavior of the needle could identify Magnetic North